// src/pages/Admin/academicManagement/CreateAcademicDepartment.tsx
import { Button, Col, Flex } from "antd";
import { FieldValues, SubmitHandler } from "react-hook-form";
import { toast } from "sonner";
import PHForm from "../../../components/form/PHForm";
import PHInput from "../../../components/form/PHInput";
import PHSelect from "../../../components/form/PHSelect";
import {
  useAddAcademicDepartmentMutation,
  useGetAcademicFacultiesQuery,
} from "../../../redux/features/admin/academicManagement.api";
import { TAcademicFaculty } from "../../../types";
import { TResponse } from "../../../types/global";

const CreateAcademicDepartment = () => {
  const [addAcademicDepartment] = useAddAcademicDepartmentMutation();
  const { data: faculties } = useGetAcademicFacultiesQuery(undefined);

  const onSubmit: SubmitHandler<FieldValues> = async (data) => {
    const toastId = toast.loading("Creating...");
    try {
      const res = (await addAcademicDepartment(data)) as TResponse<any>;
      if (res.error) {
        toast.error(res.error.data.message, { id: toastId });
      } else {
        toast.success(res?.data?.message, { id: toastId });
      }
    } catch (error) {
      toast.error("Something went wrong", { id: toastId });
    }
  };

  const facultyOptions = faculties?.data?.map((item: TAcademicFaculty) => ({
    value: item._id,
    label: item.name,
  }));

  return (
    <Flex justify="center" align="center">
      <Col span={8}>
        <PHForm onSubmit={onSubmit} formTitle="Create academic department form">
          <PHInput
            placeholder="Enter name"
            label="Name"
            name="name"
            type="text"
          />

          <PHSelect
            placeholder="Select faculty"
            label="Faculty"
            name="academicFaculty"
            options={facultyOptions}
          />

          <Button htmlType="submit">Submit</Button>
        </PHForm>
      </Col>
    </Flex>
  );
};

export default CreateAcademicDepartment;
